updates slacbot
