+++ scripts/slackbot/utils/generateGreeting.ts
@@ -1,0 +1,46 @@
+import { sample } from 'lodash';
+
+export function generateGreeting(length: number) {
+  const EMOJIS = [
+    ':wave:',
+    ':lefty_wave:',
+    ':blob-wave:',
+    ':blob-dance:',
+    ':blob_excited:',
+    ':celebrate:',
+    ':tada:',
+    ':sparkles:',
+    ':rocket:',
+    ':boom:',
+    ':zap:',
+    ':fire:',
+    ':seedling:',
+    ':herb:',
+    ':leaves:',
+    ':leafy_green:',
+  ];
+
+  const NYCTime = new Date().getUTCHours() - 5;
+  const timeOfDay = getTimeOfDay(NYCTime);
+  const GREETINGS = [
+    'Hey there!',
+    'Hello there!',
+    'Hi there!',
+    `Good ${timeOfDay}!`,
+  ];
+
+  const emoji = sample(EMOJIS);
+  const greeting = sample(GREETINGS);
+  const intro =
+    length > 1
+      ? 'Some new LeafyGreen updates today'
+      : `Just one LeafyGreen update this ${timeOfDay}`;
+
+  return `${emoji} ${greeting} ${intro}`;
+}
+
+function getTimeOfDay(time: number) {
+  if (time <= 12) return 'morning';
+  else if (time <= 18) return 'afternoon';
+  else return 'evening';
+}
